changed graphics library to position and sizes instead of separate variable (it was a ton of work)

+++ client/menus/background_rect.rs
@@ -17,11 +17,11 @@ as the height of the window. You can also retrieve the back rectangle's
 container.
  */
 pub trait BackgroundRect: Background {
-    fn set_back_rect_width(&mut self, width: i32);
+    fn set_back_rect_width(&mut self, width: f32);
     fn get_back_rect_width(
         &self,
         graphics: &GraphicsContext,
         parent_container: Option<&Container>,
-    ) -> i32;
+    ) -> f32;
     fn get_back_rect_container(&self) -> &Container;
 }
